move _credential.py up one level
