+++ sdk/search/azure-search-documents/azure/search/documents/_credential.py
