+++ sdk/search/azure-search-documents/azure/search/documents/_index/__init__.py
@@ -2,7 +2,6 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 # ------------------------------------
-from ._credential import SearchApiKeyCredential  # pylint: disable=unused-import
 from ._index_documents_batch import IndexDocumentsBatch  # pylint: disable=unused-import
 from ._search_index_client import (  # pylint: disable=unused-import
     odata,
