+++ sdk/search/azure-search-documents/azure/search/documents/_index/aio/_search_index_client_async.py
@@ -19,7 +19,7 @@ from ..._version import VERSION
 if TYPE_CHECKING:
     # pylint:disable=unused-import,ungrouped-imports
     from typing import Any, Union
-    from .._credential import SearchApiKeyCredential
+    from ... import SearchApiKeyCredential
 
 
 class SearchIndexClient(object):
